third_party/blink/web_tests/platform/generic/external/wpt/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-pointer-insert-br-expected.txt

This is a testharness.js-based test.
FAIL pointer updates (adding br elements) assert_equals: error events on b expected 1 but got 0
Harness: the test ran to completion.
